Probleem 48 opgelost.

+++ org.schoen.ben.euler/src/org/schoen/ben/euler/Problem047.java
@@ -61,7 +61,7 @@ public class Problem047 extends AbstractEulerProblem {
 			return false;
 		int primes = 0;
 		for(Integer i : factors) {
-			if(PrimeUtil.isPrime(i))
+			if(PrimeUtil.isPrime(i.longValue()))
 				primes++;
 		}
 		return primes == 4;
